squash! Fix Python 3 issue in HBB tutorial
Change spa.State to EnsembleArray and use constant number of neurons per
dimensions (this seems more sensible to me). This also changes the label
of the plot from “state” to “cleanup” which should make it easier to
figure out what is plotted (without relying on the comments at the top
of the file).